Source code was synced with Neon game development kit 4.2.3

+++ SWGF.cpp
@@ -2589,32 +2589,32 @@ Transformation::~Transformation()
 
 }
 
-void Transformation::initialize(const unsigned long int screen_width,const unsigned long int screen_height,const unsigned long int surface_width,const unsigned long int surface_height)
+void Transformation::initialize(const float screen_width,const float screen_height,const float surface_width,const float surface_height)
 {
- screen_x_factor=(float)screen_width/(float)surface_width;
- screen_y_factor=(float)screen_height/(float)surface_height;
- surface_x_factor=(float)surface_width/(float)screen_width;
- surface_y_factor=(float)surface_height/(float)screen_height;
+ screen_x_factor=screen_width/surface_width;
+ screen_y_factor=screen_height/surface_height;
+ surface_x_factor=surface_width/screen_width;
+ surface_y_factor=surface_height/screen_height;
 }
 
-unsigned long int Transformation::get_screen_x(const unsigned long int surface_x)
+float Transformation::get_screen_x(const float surface_x)
 {
- return screen_x_factor*(float)surface_x;
+ return screen_x_factor*surface_x;
 }
 
-unsigned long int Transformation::get_screen_y(const unsigned long int surface_y)
+float Transformation::get_screen_y(const float surface_y)
 {
- return screen_y_factor*(float)surface_y;
+ return screen_y_factor*surface_y;
 }
 
-unsigned long int Transformation::get_surface_x(const unsigned long int screen_x)
+float Transformation::get_surface_x(const float screen_x)
 {
- return surface_x_factor*(float)screen_x;
+ return surface_x_factor*screen_x;
 }
 
-unsigned long int Transformation::get_surface_y(const unsigned long int screen_y)
+float Transformation::get_surface_y(const float screen_y)
 {
- return surface_y_factor*(float)screen_y;
+ return surface_y_factor*screen_y;
 }
 
 Collision_Box Collision::generate_box(const unsigned long int x,const unsigned long int y,const unsigned long int width,const unsigned long int height)
